update some config file

+++ trunk/WebServer/src/com/pitaya/bookingnow/dao/AccountDao.java
@@ -5,7 +5,7 @@ import org.springframework.dao.DataAccessException;
 public interface AccountDao<T> {
 	public boolean addAccount(T entity) throws DataAccessException;
 
-	public T getAccount(Integer id) throws DataAccessException;
+	public T getAccount(Long id) throws DataAccessException;
 
 	public List<T> getList() throws DataAccessException;
 }
